Create main.c
This code is one solution to the problem CountFactors. Given a positive integer N, we must return the total number of factors for the given integer. So all we need to do is to traverse from 1 <= N and keep counting the numbers which can return remainder 0.
Check for perfect squares added.